% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abseq.R
\docType{class}
\name{AbSeq-class}
\alias{AbSeq-class}
\alias{AbSeq}
\title{AbSeq S4 class}
\value{
none
}
\description{
AbSeq S4 class
}
\section{Slots}{

\describe{
\item{\code{outputDirectory}}{character type. Output directory of AbSeq analysis}
}}

