% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqCRep.R
\docType{class}
\name{AbSeqCRep-class}
\alias{AbSeqCRep-class}
\alias{AbSeqCRep}
\title{AbSeqCompositeRepertoire}
\value{
AbSeqCRep
}
\description{
AbSeqCRep is a collection of \linkS4class{AbSeqRep} S4 objects
}
\section{Slots}{

\describe{
\item{\code{repertoires}}{list.}
}}

\examples{
# Use example data from abseqR as abseqPy's output, substitute this
# with your own abseqPy output directory
abseqPyOutput <- tempdir()
file.copy(system.file("extdata", "ex", package = "abseqR"), abseqPyOutput, recursive=TRUE)
samples <- abseqReport(file.path(abseqPyOutput, "ex"), report = 0)

# The provided example data has PCR1, PCR2, and PCR3 samples contained within
# pcr12 and pcr13 are instances of AbSeqCRep
pcr12 <- samples[["PCR1"]] + samples[["PCR2"]]
pcr13 <- samples[["PCR1"]] + samples[["PCR3"]]

# all_S is also an instance of AbSeqCRep
all_S <- pcr12 + pcr13
}
\seealso{
\linkS4class{AbSeqRep}
}
