% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abSeqPlot.R
\name{abSeqPlot}
\alias{abSeqPlot}
\title{Plots all samples (and comparisons) as specified in abseq.cfg}
\usage{
abSeqPlot(root, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{root}{string type. Root directory (-o / --outdir in AbSeq)}

\item{BPPARAM}{BiocParallel backend}
}
\value{
list type. List of Repertoire objects
}
\description{
Plots all samples (and comparisons) as specified in abseq.cfg
}
\examples{
todo
}
