% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-abseq.R
\name{abSeqPlotter}
\alias{abSeqPlotter}
\title{Monolith plotting function}
\usage{
abSeqPlotter(outputdir)
}
\arguments{
\item{outputdir}{string type. Path to abseq output directory}
}
\value{
None
}
\description{
Monolith plotting function
}
