% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAnalysis.R
\name{abundanceAnalysis}
\alias{abundanceAnalysis}
\title{Title}
\usage{
abundanceAnalysis(abundanceDirectories, abunOut, sampleNames, combinedNames,
  mashedNames)
}
\arguments{
\item{mashedNames}{}
}
\description{
Title
}
