% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeRepertoire.R
\docType{class}
\name{CompositeRepertoire-class}
\alias{CompositeRepertoire-class}
\alias{CompositeRepertoire}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{repertoires}}{list.}
}}

