% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAnalysis.R
\name{.abundanceAnalysis}
\alias{.abundanceAnalysis}
\title{Conducts abundance analysis}
\usage{
.abundanceAnalysis(abundanceDirectories, abunOut, sampleNames,
  combinedNames, mashedNames, skipDgene = FALSE, .save = TRUE)
}
\arguments{
\item{abundanceDirectories}{list type. List of sample directories}

\item{abunOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 correspondence with abundanceDirectories}

\item{combinedNames}{string type. Title "combined" sample names}

\item{mashedNames}{string type. File "mashed" names - avoid special chars}

\item{skipDgene}{logical type. Skip D gene plots?}

\item{.save}{logical type. Save ggplot as Rdata}
}
\value{
None
}
\description{
Conducts abundance analysis
}
