% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAnalysis.R
\name{.abundancePlot}
\alias{.abundancePlot}
\title{Abundance distribution}
\usage{
.abundancePlot(files, sampleNames, outputDir, skipDgene = FALSE,
  .save = TRUE)
}
\arguments{
\item{files}{list type. list of files in abundance directory}

\item{sampleNames}{vector type. 1-1 correspondance to files}

\item{outputDir}{string type.}

\item{skipDgene}{logical type. Skip D germline abundance plot if TRUE.}

\item{.save}{logical type. Save Rdata ggplot item}
}
\value{
None
}
\description{
Abundance distribution
}
