% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAnalysis.R
\name{.alignQualityHeatMaps}
\alias{.alignQualityHeatMaps}
\title{Plots all 5 alignment quality heatmaps}
\usage{
.alignQualityHeatMaps(abundanceDirectory, sampleName)
}
\arguments{
\item{abundanceDirectory}{character type.
fully qualified path to abundance directory}

\item{sampleName}{character type. sample name}
}
\value{
list of ggplotly heatmaps
}
\description{
Plots alignment quality vs:
\itemize{
  \item{mismatches}
  \item{gaps}
  \item{bitscore}
  \item{percent identity}
  \item{subject start}
}
}
