% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.allPrimerNames}
\alias{.allPrimerNames}
\title{Collect primer names from FASTA}
\usage{
.allPrimerNames(primerFile)
}
\arguments{
\item{primerFile}{string type. Path to primer file}
}
\value{
vector of primer names as seen in primerFile
}
\description{
Collect primer names from FASTA
}
