% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.aminoAcidBar}
\alias{.aminoAcidBar}
\title{Plots amino acid composition logo}
\usage{
.aminoAcidBar(df, scale, region, germ = "")
}
\arguments{
\item{df}{dataframe}

\item{scale}{logical. scale to proportion?}

\item{region}{string. which region is this}

\item{germ}{string. V germline family}
}
\value{
ggplot2 object
}
\description{
Plots amino acid composition logo
}
