% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.aminoAcidPlot}
\alias{.aminoAcidPlot}
\title{Composition logo plot}
\usage{
.aminoAcidPlot(compositionDirectory, outdir, sampleName,
  regions = c("FR1", "CDR1", "FR2", "CDR2", "FR3", "CDR3", "FR4"),
  .save = T)
}
\arguments{
\item{compositionDirectory}{string type.}

\item{outdir}{string type.}

\item{sampleName}{string type.}

\item{regions}{logical type. vector of FR/CDR regions to plot}

\item{.save}{logical type. save ggplot object}
}
\value{
none
}
\description{
Composition logo plot
}
