% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotationAnalysis.R
\name{.annotAnalysis}
\alias{.annotAnalysis}
\title{Annotation analysis}
\usage{
.annotAnalysis(annotDirectories, annotOut, sampleNames, mashedNames,
  .save = TRUE)
}
\arguments{
\item{annotDirectories}{list type. List of sample directories}

\item{annotOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 with annotDirectories}

\item{mashedNames}{string type. File output "mashed" sample names}

\item{.save}{logical type. Saves ggplot object}
}
\value{
none
}
\description{
Annotation analysis
}
