% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoireAlignLen}
\alias{.asRepertoireAlignLen}
\title{Accessor for \code{alignlen} slot}
\usage{
.asRepertoireAlignLen(object, collapse = " - ")
}
\arguments{
\item{object}{AbSeqRep object}

\item{collapse}{character type, collapse the range using this string.}
}
\value{
character type. If collapse is a string, then the ranges
are represented as `start - end` in a string, if collapse is NULL,
returns a character vector of length
two, denoting the start and end value respectively.
}
\description{
Accessor for \code{alignlen} slot
}
