% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoireChain}
\alias{.asRepertoireChain}
\title{Accessor for \code{chain} slot}
\usage{
.asRepertoireChain(object)
}
\arguments{
\item{object}{AbSeqRep object}
}
\value{
character type, the chain type of this sample
}
\description{
Accessor for \code{chain} slot
}
