% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoireDir}
\alias{.asRepertoireDir}
\title{Accessor for the \code{outdir} slot}
\usage{
.asRepertoireDir(object)
}
\arguments{
\item{object}{AbSeqRep object}
}
\value{
character type, the output directory of this object
}
\description{
Accessor for the \code{outdir} slot
}
