% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoireList}
\alias{.asRepertoireList}
\title{Accessor for \linkS4class{AbSeqCRep}'s list of \linkS4class{AbSeqRep} objects}
\usage{
.asRepertoireList(object)
}
\arguments{
\item{object}{AbSeqCRep object}
}
\value{
list type, list of \linkS4class{AbSeqRep} objects that together,
compose this \linkS4class{AbSeqCRep} object.
}
\description{
Accessor for \linkS4class{AbSeqCRep}'s list of \linkS4class{AbSeqRep} objects
}
