% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoireName}
\alias{.asRepertoireName}
\title{Accessor for the \code{name} slot}
\usage{
.asRepertoireName(object)
}
\arguments{
\item{object}{AbSeqRep object}
}
\value{
character type, the sample name of this object.
}
\description{
Accessor for the \code{name} slot
}
