% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoirePrimer3}
\alias{.asRepertoirePrimer3}
\title{Accessor for the \code{primer3end} slot}
\usage{
.asRepertoirePrimer3(object)
}
\arguments{
\item{object}{AbSeqRep object}
}
\value{
character type, the FASTA file name for primer 3' end sequences
}
\description{
Accessor for the \code{primer3end} slot
}
