% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoirePrimer5}
\alias{.asRepertoirePrimer5}
\title{Accessor for the \code{primer5end} slot}
\usage{
.asRepertoirePrimer5(object)
}
\arguments{
\item{object}{AbSeqRep object}
}
\value{
character type, the FASTA file name for primer 5' end sequences
}
\description{
Accessor for the \code{primer5end} slot
}
