% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessors-AbSeq.R
\name{.asRepertoireUpstream}
\alias{.asRepertoireUpstream}
\title{Accessor for the \code{upstream} slot}
\usage{
.asRepertoireUpstream(object)
}
\arguments{
\item{object}{AbSeqRep object}
}
\value{
character type
}
\description{
Accessor for the \code{upstream} slot
}
