% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.boxPlot}
\alias{.boxPlot}
\title{Creates a box plot}
\usage{
.boxPlot(dataframes, sampleNames, plotTitle, xlabel = "", ylabel = "",
  subs = "")
}
\arguments{
\item{dataframes}{list type. List of sample dataframes}

\item{sampleNames}{vector type. 1-1 with dataframes}

\item{plotTitle}{string type}

\item{xlabel}{string type}

\item{ylabel}{string type}

\item{subs}{string type}
}
\value{
ggplot2 object
}
\description{
Creates a box plot
}
