% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.calculateDiversityEstimates}
\alias{.calculateDiversityEstimates}
\title{Calculates Lower Bound Estimates for unseen species and Common Diversity
Indices from clonotype tables}
\usage{
.calculateDiversityEstimates(diversityDirectories, diversityOut,
  sampleNames)
}
\arguments{
\item{diversityDirectories}{list type. List of directories to diversity dir}

\item{diversityOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 with diversityDirectories sample names}
}
\value{
None
}
\description{
Employ common techniques to calculate LBE for unseen species
and commonly used diversity indices
}
