% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.canonicalizeTitle}
\alias{.canonicalizeTitle}
\title{Convert file names to human friendly text}
\usage{
.canonicalizeTitle(str)
}
\arguments{
\item{str}{string type}
}
\value{
string
}
\description{
Convert file names to human friendly text
}
