% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.capitalize}
\alias{.capitalize}
\title{Helper function to capitalize the first letter of \code{str}}
\usage{
.capitalize(str)
}
\arguments{
\item{str}{string type}
}
\value{
string, \code{str} capitalized
}
\description{
Helper function to capitalize the first letter of \code{str}
}
