% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.checkVert}
\alias{.checkVert}
\title{Checks if abseqPy has a metadata line that suggests
the orientation}
\usage{
.checkVert(filename)
}
\arguments{
\item{filename}{csv filename}
}
\value{
True if CSV metadata says "plot vertically"
}
\description{
Checks if abseqPy has a metadata line that suggests
the orientation
}
