% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.cloneDistHist}
\alias{.cloneDistHist}
\title{Marginal histogram of clonotypes (blue for shared, grey for total). The y
axis is scaled by sqrt (but it doesn't really matter anyway, since we're
stripping away the y-ticks)}
\usage{
.cloneDistHist(df.original, otherClones, lim.min, flip)
}
\arguments{
\item{df.original}{dataframe with all clones}

\item{otherClones}{clones from the other dataframe}

\item{lim.min}{x-axis minimum limit}

\item{flip}{logical type}
}
\value{
ggplot2 object
}
\description{
Marginal histogram of clonotypes (blue for shared, grey for total). The y
axis is scaled by sqrt (but it doesn't really matter anyway, since we're
stripping away the y-ticks)
}
