% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.cloneDistMarginal}
\alias{.cloneDistMarginal}
\title{Marginal density graph of clonotypes (blue for shared, grey for total, purple
for exclusive clones)}
\usage{
.cloneDistMarginal(df.original, otherClones, lim.min, flip)
}
\arguments{
\item{df.original}{dataframe with all clones}

\item{otherClones}{clones from the other dataframe}

\item{lim.min}{x-axis minimum limit}

\item{flip}{logical type}
}
\value{
ggplot2 object
}
\description{
Marginal density graph of clonotypes (blue for shared, grey for total, purple
for exclusive clones)
}
