% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.clonotypeAnalysis}
\alias{.clonotypeAnalysis}
\title{Comprehensive clonotype analyses}
\usage{
.clonotypeAnalysis(diversityDirectories, clonotypeOut, sampleNames,
  mashedNames, .save = T)
}
\arguments{
\item{diversityDirectories}{list type. List of directories to diversity dir}

\item{clonotypeOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 with diversityDirectories}

\item{mashedNames}{string type. Prefix for ooutput files using "mashed-up"}

\item{.save}{logical type. Save ggplot object?}
}
\value{
Nothing
}
\description{
Comprehensive clonotype analyses
}
