% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abseqReport.R
\name{.collateReports}
\alias{.collateReports}
\title{Collate all HTML reports into a single directory and cretate an entry
\code{index.html} file that redirects to all collated HTML files}
\usage{
.collateReports(reports, individualSamples, outputDirectory)
}
\arguments{
\item{reports}{list/vector type. Collection of strings that are path(s)
to <sample>_report.html}

\item{individualSamples}{list type. list of AbSeqRep objects. Used to
extract filtering information and \% read counts.}

\item{outputDirectory}{string type. Where should the report be placed.}
}
\value{
Nothing
}
\description{
Collate all HTML reports into a single directory and cretate an entry
\code{index.html} file that redirects to all collated HTML files
}
