% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.commonPrimerNames}
\alias{.commonPrimerNames}
\title{Collect the intersection of all primer names within a collection
of primer files}
\usage{
.commonPrimerNames(primerFiles)
}
\arguments{
\item{primerFiles}{list / vector type. Collection of primer files}
}
\value{
vector type. Vector of primerNames that are present in ALL
primerFiles. NULL if there's no intersection at all
}
\description{
Collect the intersection of all primer names within a collection
of primer files
}
