% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{.correlationTest}
\alias{.correlationTest}
\title{Conducts pearson and spearman correlation analysis on dataframe}
\usage{
.correlationTest(df)
}
\arguments{
\item{df}{dataframe with at least the following 2 columns:
+-----------------+
| prop.x | prop.y |
+-----------------+
|.....   | ....   |
+-----------------+
where prop.x and prop.y are normalized counts (i.e. frequencies) of the clones
They may contain 0 in a column to denote it being missing from sample x or y.}
}
\value{
named list of pearson, pearson.p, spearman, spearman.p
}
\description{
Conducts pearson and spearman correlation analysis on dataframe
}
