% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{.distanceMeasure}
\alias{.distanceMeasure}
\title{Computes the distance between pariwise samples}
\usage{
.distanceMeasure(df)
}
\arguments{
\item{df}{dataframe with at least the following 2 columns:
+-----------------+
| prop.x | prop.y |
+-----------------+
|.....   | ....   |
+-----------------+
where prop.x and prop.y are normalized counts (i.e. frequencies) of the clones
They may contain 0 in a column to denote it being missing from sample x or y.}
}
\value{
named list of bray.curtis, jaccard, and morisita.horn
}
\description{
Computes the distance between pariwise samples
}
