% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.diversityAnalysis}
\alias{.diversityAnalysis}
\title{Title Diversity analysis}
\usage{
.diversityAnalysis(diversityDirectories, diversityOut, sampleNames,
  mashedNames, .save = TRUE)
}
\arguments{
\item{diversityDirectories}{list type. List of directories to diversity dir}

\item{diversityOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 with diversityDirectories}

\item{mashedNames}{string type. Prefix for output files using "mashed-up"
sample names}

\item{.save}{logical type. Save ggplot object?}
}
\value{
None
}
\description{
Title Diversity analysis
}
