% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.findRepertoires}
\alias{.findRepertoires}
\title{Given a directory = <abseqPy_outputdir>/RESULT_DIR/, returns the directories (repositories) in
'directory'. That is, will not return any sample_vs_sample directories.
This is done by asserting that a 'repository'
must have an (analysis.params) file, and a summary.txt file.}
\usage{
.findRepertoires(directory)
}
\arguments{
\item{directory}{string. Path up until <abseqPy_outputdir>/RESULT_DIR/}
}
\value{
vector of strings that are samples in 'directory', note, this is NOT
a full path, but just the sample/repertoire name itself
}
\description{
A sample_vs_sample directory will not have these files.
}
