% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abseqReport.R
\name{.generateDelayedReport}
\alias{.generateDelayedReport}
\title{Generates report for all samples in 'compare'}
\usage{
.generateDelayedReport(root, compare, interactivePlot)
}
\arguments{
\item{root}{string, project root directory.}

\item{compare}{vector of strings, each string is a comparison defined
by the user (assumes that this value has been checked).}

\item{interactivePlot}{logical, whether or not to plot interactive plotly
plots.}
}
\value{
a named list of samples, each an AbSeqRep object found in "root"
}
\description{
This function is needed because we are delaying the generation
of reports until after all threads/processes have joined. There's currently
an issue with rmarkdown::render() in parallel execution, see:
https://github.com/rstudio/rmarkdown/issues/499
}
