% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqRep.R
\name{.generateReport}
\alias{.generateReport}
\title{Generates HTML report from \code{\linkS4class{AbSeqRep}} and
\code{\linkS4class{AbSeqCRep}} ojects}
\usage{
.generateReport(object, root, outputDir, interactivePlot = TRUE,
  .indexHTML = "#")
}
\arguments{
\item{object}{AbSeqCRep type.}

\item{root}{string type. Root directory of the sample(s)}

\item{outputDir}{string type. The path where the HTML will be generated}

\item{interactivePlot}{logical type. Interactive or not}

\item{.indexHTML}{character type. The back button will redirect to this link.
This is typically used to redirect users back to index.html page}
}
\value{
path (including HTML name) where the report (HTML file) was saved to
}
\description{
Generates HTML report from \code{\linkS4class{AbSeqRep}} and
\code{\linkS4class{AbSeqCRep}} ojects
}
