% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.getLineTypes}
\alias{.getLineTypes}
\title{Helper function to return line types by importance based on provided
CD/Fs regions}
\usage{
.getLineTypes(regions)
}
\arguments{
\item{regions}{a list/vector of strings (regions)}
}
\value{
vector of strings, each corresponding to the appropriate line type
for \code{regions}.
}
\description{
In the aesthetics of diversity plots (rarefaction, recapture,
and duplication), the line types should emphasise the most important
antibody region, they're ranked in ascending order of:
"FR4", "FR1", "FR2", "FR3", "CDR1", "CDR2", "CDR3", "V".
}
