% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.getTotal}
\alias{.getTotal}
\title{Get total number of samples (n)}
\usage{
.getTotal(filename)
}
\arguments{
\item{filename}{csv filename}
}
\value{
string, sample size.
}
\description{
Often enough, the CSV values supplied do not contain
raw counts but percentages (so this value will let us know exactly
the sample size).
}
