% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAnalysis.R
\name{.hmFromMatrix}
\alias{.hmFromMatrix}
\title{Plots a plotly heatmap from provided matrix}
\usage{
.hmFromMatrix(m, title, xlabel = "", ylabel = "")
}
\arguments{
\item{m}{matrix type}

\item{title}{character type}

\item{xlabel}{character type}

\item{ylabel}{character type}
}
\value{
list with keys: static and interactive (ggplot2 object and plotly
object respectivelyb)
}
\description{
Plots a plotly heatmap from provided matrix
}
