% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.inferAnalyzed}
\alias{.inferAnalyzed}
\title{Returns all samples found under \code{sampleDirectory}}
\usage{
.inferAnalyzed(sampleDirectory)
}
\arguments{
\item{sampleDirectory}{string, path to sample directory.}
}
\value{
un-normalized path to all samples under \code{sampleDirectory}
}
\description{
Returns all samples found under \code{sampleDirectory}
}
