% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abseqReport.R
\name{.loadSamplesFromString}
\alias{.loadSamplesFromString}
\title{Loads AbSeqCRep or AbSeqRep objects from a list of sampleNames}
\usage{
.loadSamplesFromString(sampleNames, root, warnMove = TRUE)
}
\arguments{
\item{sampleNames}{vector, singleton or otherwise}

\item{root}{string type. root directory}

\item{warnMove}{logical type. Warning message
("message" level, not "warning" level) if the directory has been moved?}
}
\value{
AbSeqRep or AbSeqCRep object depending on sampleNames
}
\description{
Loads AbSeqCRep or AbSeqRep objects from a list of sampleNames
}
