% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.pairwiseComparison}
\alias{.pairwiseComparison}
\title{Conduct all pairwise comparison analyses}
\usage{
.pairwiseComparison(dataframes, sampleNames, outputPath, .save = TRUE)
}
\arguments{
\item{dataframes}{list of dataframes}

\item{sampleNames}{1-1 vector corresponding to dataframes}

\item{outputPath}{string}

\item{.save}{logical}
}
\value{
nothing
}
\description{
Conduct all pairwise comparison analyses
}
