% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abundanceAnalysis.R
\name{.plotCirclize}
\alias{.plotCirclize}
\title{V-J association plot}
\usage{
.plotCirclize(sampleName, path, outputdir)
}
\arguments{
\item{sampleName}{string type}

\item{path}{string type. Path to _vjassoc.csv}

\item{outputdir}{string type}
}
\value{
None
}
\description{
V-J association plot
}
