% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.plotDiversityCurves}
\alias{.plotDiversityCurves}
\title{Plots rarefaction, recapture, and de-dup plots for specified \code{region}}
\usage{
.plotDiversityCurves(region, diversityDirectories, sampleNames,
  mashedNames, diversityOut, .save = TRUE)
}
\arguments{
\item{region}{string type. One of: "cdr", "cdr_v", and "fr". "cdr" means
CDR1-3, "cdr_v" means CDR3 and V only, and finally "fr" means FR1-4.}

\item{diversityDirectories}{list type. List of directories to diversity dir}

\item{sampleNames}{vector type. 1-1 with diversityDirectories}

\item{mashedNames}{string type. Prefix for output files using "mashed-up"}

\item{diversityOut}{string type. Output directory sample names}

\item{.save}{logical type. Save ggplot object?}
}
\value{
Nothing
}
\description{
Plots rarefaction, recapture, and de-dup plots for specified \code{region}
}
