% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.plotDuplication}
\alias{.plotDuplication}
\title{Duplication level plot}
\usage{
.plotDuplication(files, sampleNames, regions = c("CDR3", "V"))
}
\arguments{
\item{files}{list type. List of strings to _cdr_v_duplication.csv pathname}

\item{sampleNames}{vector type. Vector of strings each representing
sample names}

\item{regions}{vector type.
Which regions to include in the plot. Default = c("CDR3", "V")}
}
\value{
ggplot2 object
}
\description{
bins singletons, doubletons, and higher order clonotypes
into a line plot
}
