% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivityAnalysis.R
\name{.plotIGVErrors}
\alias{.plotIGVErrors}
\title{Plots the error distribution for IGV germlines}
\usage{
.plotIGVErrors(productivityDirectories, prodOut, sampleNames,
  combinedNames, mashedNames, .save = TRUE)
}
\arguments{
\item{productivityDirectories}{list type. List of directories}

\item{prodOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 with productivity directories}

\item{combinedNames}{string type. Title friendly "combined" sample names}

\item{mashedNames}{string type. File friendly "mashed-up" sample names}

\item{.save}{logical type, save Rdata?}
}
\value{
None
}
\description{
Plots the distribution of in-frame unproductive,
out-of-frame unproductive, and productive IGV germlines.
}
