% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.plotPrimerIGVStatus}
\alias{.plotPrimerIGVStatus}
\title{Plots, for a given \code{category} and \code{pend}, the \code{primer}
IGV indelled distribution in a bar plot}
\usage{
.plotPrimerIGVStatus(primer, pend, category, primerDirectories,
  sampleNames, primerOut, combinedNames, mashedNames, .save = TRUE)
}
\arguments{
\item{primer}{string, primer name}

\item{pend}{string, either 3 or 5 (primer end)}

\item{category}{string, either "all", "productive", or "outframe"}

\item{primerDirectories}{string type. Path to primer analysis directory}

\item{sampleNames}{vector type. 1-1 with primerDirectories}

\item{primerOut}{string type. output directory}

\item{combinedNames}{string type. Title friendly "combined" sample names}

\item{mashedNames}{string type. File friendly "mashed-up" sample names}

\item{.save}{logical type. Save Rdata?}
}
\value{
None
}
\description{
Plots the abundace of indelled primers relative to IGV germlines
}
