% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.plotPrimerIntegrity}
\alias{.plotPrimerIntegrity}
\title{Plots the distribution of primer integrity for a given \code{category} and
5' or 3' \code{pend}}
\usage{
.plotPrimerIntegrity(primerIntegrity, pend, category, primerDirectories,
  sampleNames, primerOut, combinedNames, mashedNames)
}
\arguments{
\item{primerIntegrity}{string. One of "stopcodon", "integrity",
"indelled", "indel_pos"}

\item{pend}{string, either 3 or 5 (primer end)}

\item{category}{string, either "all", "productive", or "outframe"}

\item{primerDirectories}{string type. Path to primer analysis directory}

\item{sampleNames}{vector type. 1-1 with primerDirectories}

\item{primerOut}{string type. output directory}

\item{combinedNames}{string type. Title friendly "combined" sample names}

\item{mashedNames}{string type. File friendly "mashed-up" sample names}
}
\value{
None
}
\description{
Plots the distribution of primer integrity for a given \code{category} and
5' or 3' \code{pend}
}
