% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.plotRarefaction}
\alias{.plotRarefaction}
\title{Rarefaction plot}
\usage{
.plotRarefaction(files, sampleNames, regions = c("CDR3", "V"))
}
\arguments{
\item{files}{list type. A list of files consisting of path to samples}

\item{sampleNames}{vector type. A vector of strings,
each being the name of samples in files}

\item{regions}{vector type. A vector of strings,
regions to be included. Defaults to c("CDR3", "V")}
}
\value{
ggplot2 object
}
\description{
Plots the number of unique clonotypes (on the y-axis)
drawn from a sample size on the x axis. The number of unique clonotypes
is averaged over 5 repeated rounds.
}
