% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.plotRecapture}
\alias{.plotRecapture}
\title{Title Plots recapture}
\usage{
.plotRecapture(files, sampleNames, regions = c("CDR3", "V"))
}
\arguments{
\item{files}{list type. List of _cdr_v_recapture.csv.gz files.}

\item{sampleNames}{vector type. A vector of strings each
representing the name of samples in files.}

\item{regions}{vector type. A vector of strings,
regions to be included in the plot. defaults to c("CDR3", "V")}
}
\value{
ggplot2 object
}
\description{
Title Plots recapture
}
