% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{.plotSamples}
\alias{.plotSamples}
\title{Monolith AbSeq Plot function}
\usage{
.plotSamples(sampleNames, directories, analysis, outputDir, primer5Files,
  primer3Files, upstreamRanges, skipDgene = FALSE)
}
\arguments{
\item{sampleNames}{vector type. Vector of sample names in strings}

\item{directories}{vector type. Vector of directories in strings, must be
1-1 with sampleNames}

\item{analysis}{vector / list type. What analysis to plot for. If sampleNames
or directories is > 1 (i.e. AbSeqCRep), then make sure that it's
an intersection of all analysis conducted by the repertoires, otherwise, it
wouldn't make sense}

\item{outputDir}{string type. Where to dump the output}

\item{primer5Files}{vector / list type. Collection of strings that the sample
used for primer5 analysis. If sample N doesn't have a primer 5 file,
leave it as anthing but a valid file path.}

\item{primer3Files}{vector / list type. Collection of strings that the sample
used for primer 3 analysis. If sample N doesn't have a primer 3 file,
leave it as anthing but a valid file path.}

\item{upstreamRanges}{list type. Collection of "None"s or vector
denoting upstreamStart and upstreamEnd for each sample.}

\item{skipDgene}{logical type. Whether or not to skip D gene distribution plot}
}
\value{
none
}
\description{
Monolith AbSeq Plot function
}
