% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{.plotSpectratype}
\alias{.plotSpectratype}
\title{Spectratype plotter}
\usage{
.plotSpectratype(dataframes, sampleNames, region, title = "Spectratype",
  subtitle = "", xlabel = "Length(AA)", ylabel = "Distribution",
  showLabel = F)
}
\arguments{
\item{dataframes}{list type. List of dataframes.}

\item{sampleNames}{vector type. 1-1 correspondance with dataframes}

\item{region}{string type. Region that will be displayed in the plot title.
This specifies which region this spectratype belongs to. If not supplied,
a default (start, end) range will be displayed instead}

\item{title}{string type. Ignored if region is specified.}

\item{subtitle}{string type}

\item{xlabel}{string type}

\item{ylabel}{string type}

\item{showLabel}{bool type. Show geom_text? - Ignored if samples > 1}
}
\value{
ggplot2 object
}
\description{
Plots length distribution
}
