% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivityAnalysis.R
\name{.productivityAnalysis}
\alias{.productivityAnalysis}
\title{Conducts productivty analysis}
\usage{
.productivityAnalysis(productivityDirectories, prodOut, sampleNames,
  combinedNames, mashedNames, .save = TRUE)
}
\arguments{
\item{productivityDirectories}{list type. List of directories}

\item{prodOut}{string type. Output directory}

\item{sampleNames}{vector type. 1-1 with productivity directories}

\item{combinedNames}{string type. Title friendly "combined" sample names}

\item{mashedNames}{string type. File friendly "mashed-up" sample names}

\item{.save}{logical type. Save Rdata}
}
\value{
None
}
\description{
Conducts productivty analysis
}
