% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivityAnalysis.R
\name{.productivityPlot}
\alias{.productivityPlot}
\title{Summary of (un)productivity}
\usage{
.productivityPlot(dataframes, sampleNames)
}
\arguments{
\item{dataframes}{list type. List of sample dataframes}

\item{sampleNames}{vector type. 1-1 with dataframes}
}
\value{
ggplot2 object
}
\description{
Summary of (un)productivity
}
