% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.readSummary}
\alias{.readSummary}
\title{Return value specifed by key from AbSeq's summary file}
\usage{
.readSummary(sampleRoot, key)
}
\arguments{
\item{sampleRoot}{sample's root directory. For example,
\code{/path/to/<outputdir>/reports/<sample_name>}.}

\item{key}{character type. Possible values are (though they might change)
\itemize{
   \item{RawReads}
   \item{AnnotatedReads}
   \item{FilteredReads}
   \item{ProductiveReads}
}}
}
\value{
value associated with key from summary file. "NA" (in string) if the field
is not available refer to util.R for the key values
}
\description{
Return value specifed by key from AbSeq's summary file
}
