% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.regionAnalysis}
\alias{.regionAnalysis}
\title{Title Shows varying regions for a given clonotype defined by its CDR3}
\usage{
.regionAnalysis(path, sampleName, top = 15)
}
\arguments{
\item{path}{string type. Path to diversity folder
where <sampleName>_clonotype_diversity_region_analysis.csv.gz is located}

\item{sampleName}{string type}

\item{top}{int type. Top N number of clones to analyze}
}
\value{
ggplot2 object
}
\description{
Title Shows varying regions for a given clonotype defined by its CDR3
}
