% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.reportLBE}
\alias{.reportLBE}
\title{Reports abundance-based (Lower bound) diversity estimates using the Vegan package}
\usage{
.reportLBE(df)
}
\arguments{
\item{df}{clonotype dataframe. Vegan format:
+---------------------------+
| S.1| S.2| S.3 | S.4 | ... |   (each species should have its own column)
+---------------------------+
| v1 |v2  | v3  | ....      |   (each species' count values are placed in the corresponding column)
+---------------------------+}
}
\value{
dataframe with the format:
+----------------------------------------------------------------+
| S.obs | S.chao1 | se.chao1 | S.ACE | se.ACE | s.jack1 | s.jack2|
+----------------------------------------------------------------+
| v1    |  v2 ....                                               |
+----------------------------------------------------------------+
}
\description{
Reports abundance-based (Lower bound) diversity estimates using the Vegan package
}
