% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.saveAs}
\alias{.saveAs}
\title{Saves ggplot object as a Rdata file.}
\usage{
.saveAs(.save, filename, plot)
}
\arguments{
\item{.save}{logical type. Whether or not we should save.}

\item{filename}{string.}

\item{plot}{ggplot object.}
}
\value{
nothing
}
\description{
It's a convinient function that does the check and saves
at the same time, for brevity within other areas of the code (to eliminate
repeated if checks).
}
