% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.scatterPlot}
\alias{.scatterPlot}
\title{Title Creates a scatter plot}
\usage{
.scatterPlot(df1, df2, name1, name2, cloneClass)
}
\arguments{
\item{df1}{dataframe for sample 1}

\item{df2}{dataframe for sample 2}

\item{name1}{string type, Sample 1 name}

\item{name2}{string type. Sample 2 name}

\item{cloneClass}{string type.
What region was used to classify clonotypes - appears in title. For example,
CDR3 or V region}
}
\value{
ggplot2 object
}
\description{
Title Creates a scatter plot
}
