% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.scatterPlotComplex}
\alias{.scatterPlotComplex}
\title{Creates a complex scatter plot}
\usage{
.scatterPlotComplex(df.union, df1, df2, name1, name2, cloneClass)
}
\arguments{
\item{df.union}{a 'lossless' dataframe created by intersecting sample1 and
sample2's dataframes. It should contain NAs where clones that appear in one
sample doesn't appear in the other. For example:

+-------------------------------------------------+
| Clonotype | prop.x | prop.y | Count.x | Count.y |
+-------------------------------------------------+
| ABCDEF       NA       0.01      NA        210   |
| ......                                          |
+-------------------------------------------------+}

\item{df1}{dataframe for sample 1}

\item{df2}{dataframe for sample 2}

\item{name1}{string type, Sample 1 name}

\item{name2}{string type. Sample 2 name}

\item{cloneClass}{string type.
What region was used to classify clonotypes - appears in title. For example,
CDR3 or V region

this plotting techique was
shamelessly plagarised from
https://github.com/mikessh/vdjtools/blob/master/src/main/resources/rscripts/intersect_pair_scatter.r
(VDJTools) with minor modifications}
}
\value{
ggplot2 object
}
\description{
Creates a complex scatter plot
}
