% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstreamAnalysis.R
\name{.secretionSignalAnalysis}
\alias{.secretionSignalAnalysis}
\title{Secretion signal analysis}
\usage{
.secretionSignalAnalysis(secDirectories, secOut, sampleNames,
  combinedNames, mashedNames, upstreamRanges, .save = TRUE)
}
\arguments{
\item{secDirectories}{list type. Secretion signal directories where files
are located}

\item{secOut}{string type. Where to dump output}

\item{sampleNames}{vector type. 1-1 with secDirectories}

\item{combinedNames}{string type. Title friendly string}

\item{mashedNames}{string type. File name friendly string}

\item{upstreamRanges}{list type. Upstream ranges for each sample.
If length(secDirectories) > 1, the plots will only be generated for
upstream ranges that are present in ALL samples. (i.e. the intersection)}

\item{.save}{logical type, save Rdata?}
}
\value{
none
}
\description{
Generates all the required plots for Secretion signal analysis.
This includes upstream length distributions and upstream sequence validity.
}
