% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{.substituteStringInFile}
\alias{.substituteStringInFile}
\title{Substitutes the first occurance of `key` with `value` in `filename`}
\usage{
.substituteStringInFile(filename, key, value, fixed = FALSE)
}
\arguments{
\item{filename}{character type}

\item{key}{character type}

\item{value}{character type}

\item{fixed}{logical type}
}
\value{
None
}
\description{
Substitutes the first occurance of `key` with `value` in `filename`
}
