% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.topNDist}
\alias{.topNDist}
\title{Title Clonotype table}
\usage{
.topNDist(dataframes, sampleNames, top = 10)
}
\arguments{
\item{dataframes}{list type. List of dataframes.}

\item{sampleNames}{vector type. vector of strings representing sample
names should have one-to-one correspondence with dataframes}

\item{top}{int type. Top N clonotypes to plot}
}
\value{
None
}
\description{
Title Clonotype table
}
