% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upstreamAnalysis.R
\name{.upstreamAnalysis}
\alias{.upstreamAnalysis}
\title{Conducts upstream analysis}
\usage{
.upstreamAnalysis(upstreamDirectories, upstreamOut, expectedLength,
  upstreamLengthRange, sampleNames, combinedNames, mashedNames, secsig)
}
\arguments{
\item{upstreamDirectories}{list type. List of sample directories}

\item{upstreamOut}{string type. Output directory}

\item{expectedLength}{int type. Expected length of upstream sequences.
Can be infinite}

\item{upstreamLengthRange}{string type. start_end format}

\item{sampleNames}{vector type. 1-1 with upstream directories}

\item{combinedNames}{string type. Title friendly "combined" sample names}

\item{mashedNames}{string type. File friendly "mashed-up" sample names}

\item{secsig}{boolean type. True if upstream is for secretion signal,
False if it's for 5UTR}
}
\value{
None
}
\description{
Conducts upstream analysis
}
