% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise.R
\name{.vennIntersection}
\alias{.vennIntersection}
\title{Title Creates Venndiagram for clonotype intersection}
\usage{
.vennIntersection(dataframes, sampleNames, outFile, top = Inf)
}
\arguments{
\item{dataframes}{list type. List of sample dataframes. Only accepts 2 - 5
samples. Warning message will be generated for anything outside of the range}

\item{sampleNames}{vector type. 1-1 with dataframes}

\item{outFile}{string type. Filename to be saved as}

\item{top}{int type. Top N cutoff, defaults to ALL clones if not specified}
}
\value{
Nothing
}
\description{
Title Creates Venndiagram for clonotype intersection
}
