% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\name{fromDir}
\alias{fromDir}
\title{Produces a named list of \linkS4class{Repertoire} object(s)}
\usage{
fromDir(directory)
}
\arguments{
\item{directory}{character type. This should be the same string as the
argument provided to \code{abseqPy}'s \code{-o} or \code{--outdir}.}
}
\value{
named list of \linkS4class{Repertoire} objects. The names of the list
is the exact string used to name each repertoire object.
}
\description{
Produces a named list of \linkS4class{Repertoire} object(s)
}
\examples{
\dontrun{
# assuming abseqPy was run with -o results (or --outdir results)
samples <- fromDir("results")
}
}
\seealso{
\linkS4class{Repertoire}
}
