% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqCRep.R
\docType{methods}
\name{+,AbSeqCRep,AbSeqCRep-method}
\alias{+,AbSeqCRep,AbSeqCRep-method}
\title{Combines 2 \linkS4class{AbSeqCRep} objects together for comparison}
\usage{
\S4method{+}{AbSeqCRep,AbSeqCRep}(e1, e2)
}
\arguments{
\item{e1}{AbSeqCRep.}

\item{e2}{AbSeqCRep.}
}
\value{
\linkS4class{AbSeqCRep} object. Calling \code{abseqR}'s
functions on this object will always result in a comparison.
}
\description{
Combines 2 \linkS4class{AbSeqCRep} objects together for comparison
}
\examples{
\dontrun{
# 'load' AbSeqRep objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results")

# assuming there are samples named "Sample1", "Sample2", "Sample3", and "Sample4"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]
S2S4 <- samples[["Sample2"]] + samples[["Sample4"]]

all.S <- S1S3 + S2S4

# generate plots and report for this new comparison
report(all.S, "s1_vs_s2_vs_s3_vs_s4")
}
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{AbSeqRep}s
}
