% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compositeRepertoire.R
\docType{methods}
\name{+,CompositeRepertoire,CompositeRepertoire-method}
\alias{+,CompositeRepertoire,CompositeRepertoire-method}
\title{Title}
\usage{
\S4method{+}{CompositeRepertoire,CompositeRepertoire}(e1, e2)
}
\arguments{
\item{e1}{CompositeRepertoire.}

\item{e2}{CompositeRepertoire.}
}
\description{
Title
}
