% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{methods}
\name{+,CompositeRepertoire,Repertoire-method}
\alias{+,CompositeRepertoire,Repertoire-method}
\title{Title}
\usage{
\S4method{+}{CompositeRepertoire,Repertoire}(e1, e2)
}
\arguments{
\item{e1}{CompositeRepertoire.}

\item{e2}{Repertoire.}
}
\description{
Title
}
