% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{methods}
\name{+,Repertoire,CompositeRepertoire-method}
\alias{+,Repertoire,CompositeRepertoire-method}
\title{Combines a \linkS4class{Repertoire} object with
a \linkS4class{CompositeRepertoire} object together for comparison}
\usage{
\S4method{+}{Repertoire,CompositeRepertoire}(e1, e2)
}
\arguments{
\item{e1}{Repertoire.}

\item{e2}{CompositeRepertoire.}
}
\value{
\linkS4class{CompositeRepertoire} object. Calling \code{abseqR}'s
functions on this object will always result in a comparison.
}
\description{
Combines a \linkS4class{Repertoire} object with
a \linkS4class{CompositeRepertoire} object together for comparison
}
\examples{
\dontrun{
# 'load' Repertoire objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results")

# assuming there are samples named "Sample1", "Sample3", and "Sample4"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]
S4 <- samples[["Sample4"]]

S4S1S3 <- S4 + S1S3

# generate plots and report for this new comparison
report(S4S1S3, "s4_vs_s1_vs_s3")
}
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{Repertoire}s
}
