% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{methods}
\name{+,Repertoire,Repertoire-method}
\alias{+,Repertoire,Repertoire-method}
\title{Title}
\usage{
\S4method{+}{Repertoire,Repertoire}(e1, e2)
}
\arguments{
\item{e1}{Repertoire.}

\item{e2}{Repertoire.}
}
\description{
Title
}
