% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{class}
\name{Repertoire-class}
\alias{Repertoire-class}
\alias{Repertoire}
\title{Title}
\description{
Title
}
\section{Slots}{

\describe{
\item{\code{f1}}{character.}

\item{\code{f2}}{character.}

\item{\code{chain}}{character.}

\item{\code{task}}{character.}

\item{\code{name}}{character.}

\item{\code{bitscore}}{numeric.}

\item{\code{qstart}}{numeric.}

\item{\code{sstart}}{numeric.}

\item{\code{alignlen}}{numeric.}

\item{\code{clonelimit}}{numeric.}

\item{\code{log}}{character.}

\item{\code{merger}}{character.}

\item{\code{fmt}}{character.}

\item{\code{sites}}{character.}

\item{\code{primer5end}}{ANY.}

\item{\code{primer3end}}{ANY.}

\item{\code{trim5}}{numeric.}

\item{\code{trim3}}{numeric.}

\item{\code{outdir}}{character.}

\item{\code{primer5endoffset}}{numeric.}

\item{\code{threads}}{numeric.}

\item{\code{upstream}}{character.}

\item{\code{seqtype}}{character.}

\item{\code{database}}{character.}

\item{\code{actualqstart}}{numeric.}

\item{\code{fr4cut}}{logical.}

\item{\code{domainSystem}}{character.}

\item{\code{primer}}{numeric.}
}}

