% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqRep.R
\name{report}
\alias{report}
\title{Plots \linkS4class{AbSeqRep} or
\linkS4class{AbSeqCRep} object to the specfied directory}
\usage{
report(object, outputDir, report = 3)
}
\arguments{
\item{object}{AbSeqRep or AbSeqCRep object to plot}

\item{outputDir}{string type. where to save the files to}

\item{report}{integer type. The possible values are:
\itemize{
  \item{0 - does nothing (returns named list of \linkS4class{AbSeqRep} object
  , taken from the first parameter). This option isn't useful, but is
  preserved here for the sake of consistency with \code{\link{abseqReport}}}
  \item{1 - generates plots for csv files}
  \item{2 - generates a report that collates all plots}
  \item{3 - generates interactive plots in report}
}
each value also does what the previous values do. For example, \code{report = 2}
will return a named list of \linkS4class{AbSeqRep} objects, plot csv files,
and generate a (non-interactive)HTML report that collates all the plots together.}
}
\value{
named list. List of \linkS4class{AbSeqRep} objects. The names of
the list are taken directly from the repertoire object itself. This return
value is consistent with the return value of \code{\link{abseqReport}}
}
\description{
This method is analogous to \code{\link{abseqReport}}.
The only difference is that this method accepts \linkS4class{AbSeqRep} or
\linkS4class{AbSeqCRep} objects as its first parameter, and the
\code{outputDir} specifies where to store the result.
}
\examples{
\dontrun{
# 'load' AbSeqRep objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results")

# assuming there are samples named "Sample1", "Sample2", "Sample3", and "Sample4"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]
S2S4 <- samples[["Sample2"]] + samples[["Sample4"]]

all.S <- S1S3 + S2S4

# generate plots and report for this new comparison
report(all.S, "s1_vs_s2_vs_s3_vs_s4")

# generate plots only
report(all.S, "s1_vs_s2_vs_s3_vs_s4", report = 1)

# generate plots, and a non-interactive report
report(all.S, "s1_vs_s2_vs_s3_vs_s4", report = 2)

# generate plots, and an interactive report
report(all.S, "s1_vs_s2_vs_s3_vs_s4", report = 3)
}
}
\seealso{
\code{\link{abseqReport}}. Analogus function, but takes input from
a string that signifies the output directory (similar to the argument supplied to abseqPy).

\linkS4class{AbSeqRep}

\linkS4class{AbSeqCRep}
}
