% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqCRep.R
\docType{class}
\name{AbSeqCRep-class}
\alias{AbSeqCRep-class}
\alias{AbSeqCRep}
\title{AbSeqCompositeRepertoire}
\value{
AbSeqCRep
}
\description{
AbSeqCRep is a collection of \linkS4class{AbSeqRep} S4 objects
}
\section{Slots}{

\describe{
\item{\code{repertoires}}{list.}
}}

\examples{
\dontrun{
# 'load' AbSeqRep objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results", report = 0)

# assuming there are samples named "Sample1", "Sample2", "Sample3", and "Sample4"

# S1S3 is an instance of AbSeqCRep
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]

# S2S4 is an instance of AbSeqCRep
S2S4 <- samples[["Sample2"]] + samples[["Sample4"]]

# all.S is an instance of AbSeqCRep
all.S <- S1S3 + S2S4
}
}
\seealso{
\linkS4class{AbSeqRep}
}
