% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abSeqPlot.R
\name{abSeqPlot}
\alias{abSeqPlot}
\title{Plots all samples and comparisons as specified in \code{abseq.cfg}}
\usage{
abSeqPlot(root, BPPARAM = BiocParallel::bpparam())
}
\arguments{
\item{root}{string type. Root directory as specified in
\code{-o} or \code{--outdir} in AbSeqPy.}

\item{BPPARAM}{BiocParallel backend. Configures the parallel implementation.
Refer to \href{https://bioconductor.org/packages/release/bioc/html/BiocParallel.html}{BiocParallel}
for more information.}
}
\value{
list type. List of \code{\linkS4class{Repertoire}} objects
}
\description{
Plots all samples and comparisons as specified in \code{abseq.cfg}
}
\examples{
todo
}
\seealso{
\code{\linkS4class{Repertoire}}
}
