% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.canonicalizeTitle}
\alias{.canonicalizeTitle}
\title{Convert short-form names to human friendly speech}
\usage{
.canonicalizeTitle(str)
}
\arguments{
\item{str}{string type}
}
\value{
string
}
\description{
Convert short-form names to human friendly speech
}
