% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diversityAnalysis.R
\name{.plotRecapture}
\alias{.plotRecapture}
\title{Plots capture-recapture}
\usage{
.plotRecapture(files, sampleNames, regions = c("CDR3", "V"))
}
\arguments{
\item{files}{list type. List of _cdr_v_recapture.csv.gz files.}

\item{sampleNames}{vector type. A vector of strings each
representing the name of samples in files.}

\item{regions}{vector type. A vector of strings,
regions to be included in the plot. defaults to c("CDR3", "V")}
}
\value{
ggplot2 object
}
\description{
Plots the percent of recapture clonotypes (on the y-axis)
drawn from a repeated (with replacement) sample size on the x axis. The
percentage of recaptured clonotypes is averaged over 5 recapture rounds.
}
