% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerAnalysis.R
\name{.primerAnalysis}
\alias{.primerAnalysis}
\title{Conducts primer specificity analysis}
\usage{
.primerAnalysis(primerDirectories, primer5Files, primer3Files, primerOut,
  sampleNames, combinedNames, mashedNames, .save = TRUE)
}
\arguments{
\item{primerDirectories}{string type. Path to primer analysis directory}

\item{primer5Files}{vector / list type. 5' end primer files}

\item{primer3Files}{vector / list type. 3' end primer files}

\item{primerOut}{string type. output directory}

\item{sampleNames}{vector type. 1-1 with primerDirectories}

\item{combinedNames}{string type. Title friendly "combined" sample names}

\item{mashedNames}{string type. File friendly "mashed-up" sample names}

\item{.save}{logical type. Save Rdata?}
}
\value{
None
}
\description{
Conducts primer specificity analysis
}
