% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivityAnalysis.R
\name{.prodDistPlot}
\alias{.prodDistPlot}
\title{Plots a distribution plot for different productivity analysis files}
\usage{
.prodDistPlot(productivityDirectories, sampleNames, title, reg, saveNames,
  regions = c("cdr1", "cdr2", "cdr3", "fr1", "fr2", "fr3", "igv", "igd",
  "igj"), .save = TRUE)
}
\arguments{
\item{productivityDirectories}{vector type.
directories where all productivity csv
files lives (usually <samplename>/productivity/)}

\item{sampleNames}{vector type.}

\item{title}{string type.}

\item{reg}{string type.
Regular expression to find the right files for this particular distribution plot (see samples in masterScript.R)}

\item{saveNames}{vector type.
Vector of file names to save in the order of regions}

\item{regions}{string type.
Most of the dist plots are regional based. use c("") if no regions are involved}

\item{.save}{logical type. Save Rdata}
}
\value{
None
}
\description{
Plots a distribution plot for different productivity analysis files
}
