% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/productivityAnalysis.R
\name{.productivityPlot}
\alias{.productivityPlot}
\title{Summary of productivity}
\usage{
.productivityPlot(dataframes, sampleNames)
}
\arguments{
\item{dataframes}{list type. List of sample dataframes}

\item{sampleNames}{vector type. 1-1 with dataframes}
}
\value{
ggplot2 object
}
\description{
Shows the percentage of 1. productivity, 2. non-functional +
reason for being unproductive,
i.e. "Stop Codon" or "Out of frame" or "Stop & Out"
}
