% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqRep.R
\docType{methods}
\name{+,AbSeqCRep,AbSeqRep-method}
\alias{+,AbSeqCRep,AbSeqRep-method}
\title{Combines a \linkS4class{AbSeqCRep} object with
a \linkS4class{AbSeqRep} object together for comparison}
\usage{
\S4method{+}{AbSeqCRep,AbSeqRep}(e1, e2)
}
\arguments{
\item{e1}{AbSeqCRep.}

\item{e2}{AbSeqRep.}
}
\value{
\linkS4class{AbSeqCRep} object. Calling \code{abseqR}'s
functions on this object will always result in a comparison.
}
\description{
Combines a \linkS4class{AbSeqCRep} object with
a \linkS4class{AbSeqRep} object together for comparison
}
\examples{
\dontrun{
# 'load' AbSeqRep objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results", report = 0)

# assuming there are samples named "Sample1", "Sample3", and "Sample4"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]
S4 <- samples[["Sample4"]]

S1S3S4 <- S1S3 + S4

# generate plots and report for this new comparison
report(S1S3S4, "s1_vs_s3_vs_s4")
}
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{AbSeqRep}s
}
