% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqRep.R
\docType{methods}
\name{+,AbSeqRep,AbSeqRep-method}
\alias{+,AbSeqRep,AbSeqRep-method}
\title{Combines 2 \linkS4class{AbSeqRep} objects together for comparison}
\usage{
\S4method{+}{AbSeqRep,AbSeqRep}(e1, e2)
}
\arguments{
\item{e1}{AbSeqRep object.}

\item{e2}{AbSeqRep object.}
}
\value{
\linkS4class{AbSeqCRep} object. Calling \code{abseqR}'s
functions on this object will always result in a comparison.
}
\description{
Combines 2 \linkS4class{AbSeqRep} objects together for comparison
}
\examples{
\dontrun{
# 'load' AbSeqRep objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results")

# assuming there are samples named "Sample1" and "Sample3"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]

# generate plots and report for this new comparison
report(S1S3, "s1_vs_s3")
}
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{AbSeqRep}s
}
