% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{methods}
\name{+,CompositeRepertoire,Repertoire-method}
\alias{+,CompositeRepertoire,Repertoire-method}
\title{Combines a \linkS4class{CompositeRepertoire} object with
a \linkS4class{Repertoire} object together for comparison}
\usage{
\S4method{+}{CompositeRepertoire,Repertoire}(e1, e2)
}
\arguments{
\item{e1}{CompositeRepertoire.}

\item{e2}{Repertoire.}
}
\value{
\linkS4class{CompositeRepertoire} object. Calling \code{abseqR}'s
functions on this object will always result in a comparison.
}
\description{
Combines a \linkS4class{CompositeRepertoire} object with
a \linkS4class{Repertoire} object together for comparison
}
\examples{
\dontrun{
# 'load' Repertoire objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results")

# assuming there are samples named "Sample1", "Sample3", and "Sample4"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]
S4 <- samples[["Sample4"]]

S1S3S4 <- S1S3 + S4

# generate plots and report for this new comparison
plotRepertoires(S1S3S4, "s1_vs_s3_vs_s4")
}
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{Repertoire}s
}
