% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{methods}
\name{+,Repertoire,CompositeRepertoire-method}
\alias{+,Repertoire,CompositeRepertoire-method}
\title{Title}
\usage{
\S4method{+}{Repertoire,CompositeRepertoire}(e1, e2)
}
\arguments{
\item{e1}{Repertoire.}

\item{e2}{CompositeRepertoire.}
}
\description{
Title
}
