% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\docType{methods}
\name{+,Repertoire,Repertoire-method}
\alias{+,Repertoire,Repertoire-method}
\title{Combines 2 \linkS4class{Repertoire} objects together for comparison}
\usage{
\S4method{+}{Repertoire,Repertoire}(e1, e2)
}
\arguments{
\item{e1}{Repertoire object.}

\item{e2}{Repertoire object.}
}
\value{
\linkS4class{CompositeRepertoire} object. Calling \code{abseqR}'s
functions on this object will always result in a comparison.
}
\description{
Combines 2 \linkS4class{Repertoire} objects together for comparison
}
\examples{
\dontrun{
# 'load' Repertoire objects using abseqReport (ignoring plots, report, etc..)
# also assumes there's a result/ directory in current working directory,
# where result/ is the same argument passed to abseqPy's -o or --outdir parameter
samples <- abseqReport("results")

# assuming there are samples named "Sample1" and "Sample3"
S1S3 <- samples[["Sample1"]] + samples[["Sample3"]]

# generate plots and report for this new comparison
report(S1S3, "s1_vs_s3")
}
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{Repertoire}s
}
