% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AbSeqRep.R
\name{report}
\alias{report}
\title{Plots \linkS4class{AbSeqRep} or
\linkS4class{AbSeqCRep} object to the specfied directory}
\usage{
report(object, outputDir, report = TRUE, interactivePlot = TRUE)
}
\arguments{
\item{object}{AbSeqRep or AbSeqCRep object to plot}

\item{outputDir}{string type. where to save the files to}

\item{report}{logical type. Should HTML report be generated}

\item{interactivePlot}{logical type. Should the HTML report's plots be
interactive (using plotly)? This argument will be ignored if \code{report}
is \code{FALSE}}
}
\value{
nothing
}
\description{
Plots \linkS4class{AbSeqRep} or
\linkS4class{AbSeqCRep} object to the specfied directory
}
\examples{
todo
}
\seealso{
\code{\link{abseqReport}} returns a \code{list} of \code{AbSeqRep}s

\linkS4class{AbSeqRep}

\linkS4class{AbSeqCRep}
}
