% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repertoire.R
\name{report}
\alias{report}
\title{Plots \linkS4class{Repertoire} or
\linkS4class{CompositeRepertoire} object to the specfied directory}
\usage{
report(object, outputDir, report = TRUE, interactivePlot = TRUE)
}
\arguments{
\item{object}{repertoire or composite repertoire object to plot}

\item{outputDir}{string type. where to save the files to}

\item{report}{logical type. Should HTML report be generated}

\item{interactivePlot}{logical type. Should the HTML report's plots be
interactive (using plotly)? This argument will be ignored if \code{report}
is \code{FALSE}}
}
\value{
nothing
}
\description{
Plots \linkS4class{Repertoire} or
\linkS4class{CompositeRepertoire} object to the specfied directory
}
\examples{
todo
}
